requirements.txt
litellm[proxy]
exa_py
crewai>=0.30.0
crewai-tools>=0.4.0
yfinance>=0.2.40
pandas>=2.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
streamlit>=1.35.0
litellm>=1.35.0